+++ programmers/level1/19.예산/jungwoo.java
@@ -1,17 +1,19 @@
 import java.util.*;
-import java.lang.*;
 
 class Solution {
-    public int solution(int n) {
+    public int solution(int[] d, int budget) {
+        int answer = 0;
+        Arrays.sort(d);
+        int sum = 0;
+        for(int i : d){
+            sum += i;
+            if(sum > budget){
+                break;
+            }else{
+                answer++;
+            }
+        }
 
-        String temp = Integer.toString(n, 3);
-        String reverse = "";
-
-        // for(int i = temp.length() - 1; i >= 0; i--) {
-        //     reverse = reverse + temp.charAt(i);
-        // }
-        temp = new StringBuilder(temp).reverse().toString();
-
-        return Integer.parseInt(temp, 3);
+        return answer;
     }
 }
